component 분리, router, user 분리
