+++ src/main.js
@@ -1,322 +1,40 @@
-const getMainPage = () => {
-  const user = JSON.parse(localStorage.getItem('user'));
-  if (user) {
-    document.querySelector('#root').innerHTML = `
-<div class="bg-gray-100 min-h-screen flex justify-center">
-    <div class="max-w-md w-full">
-      <header class="bg-blue-600 text-white p-4 sticky top-0">
-        <h1 class="text-2xl font-bold">항해플러스</h1>
-      </header>
+import { Main, LoginMain } from '../components/Main';
+import Profile from '../components/Profile';
+import Login from '../components/Login';
+import Error from '../components/Error';
+import Router from '../utils/router';
+import User from '../store/User';
 
-      <nav class="bg-white shadow-md p-2 sticky top-14">
-        <ul class="flex justify-around">
-          <li><a href="/" class="text-blue-600">홈</a></li>
-          <li><a href="/profile" class="text-gray-600">프로필</a></li>
-          <li><a id="logout" href="/logout" class="text-gray-600">로그아웃</a></li>
-        </ul>
-      </nav>
-
-      <main class="p-4">
-        <div class="mb-4 bg-white rounded-lg shadow p-4">
-          <textarea class="w-full p-2 border rounded" placeholder="무슨 생각을 하고 계신가요?"></textarea>
-          <button class="mt-2 bg-blue-600 text-white px-4 py-2 rounded">게시</button>
-        </div>
-
-        <div class="space-y-4">
-
-          <div class="bg-white rounded-lg shadow p-4">
-            <div class="flex items-center mb-2">
-              <img src="https://via.placeholder.com/40" alt="프로필" class="rounded-full mr-2">
-              <div>
-                <p class="font-bold">홍길동</p>
-                <p class="text-sm text-gray-500">5분 전</p>
-              </div>
-            </div>
-            <p>오늘 날씨가 정말 좋네요. 다들 좋은 하루 보내세요!</p>
-            <div class="mt-2 flex justify-between text-gray-500">
-              <button>좋아요</button>
-              <button>댓글</button>
-              <button>공유</button>
-            </div>
-          </div>
-
-          <div class="bg-white rounded-lg shadow p-4">
-            <div class="flex items-center mb-2">
-              <img src="https://via.placeholder.com/40" alt="프로필" class="rounded-full mr-2">
-              <div>
-                <p class="font-bold">김철수</p>
-                <p class="text-sm text-gray-500">15분 전</p>
-              </div>
-            </div>
-            <p>새로운 프로젝트를 시작했어요. 열심히 코딩 중입니다!</p>
-            <div class="mt-2 flex justify-between text-gray-500">
-              <button>좋아요</button>
-              <button>댓글</button>
-              <button>공유</button>
-            </div>
-          </div>
-
-          <div class="bg-white rounded-lg shadow p-4">
-            <div class="flex items-center mb-2">
-              <img src="https://via.placeholder.com/40" alt="프로필" class="rounded-full mr-2">
-              <div>
-                <p class="font-bold">이영희</p>
-                <p class="text-sm text-gray-500">30분 전</p>
-              </div>
-            </div>
-            <p>오늘 점심 메뉴 추천 받습니다. 뭐가 좋을까요?</p>
-            <div class="mt-2 flex justify-between text-gray-500">
-              <button>좋아요</button>
-              <button>댓글</button>
-              <button>공유</button>
-            </div>
-          </div>
-
-          <div class="bg-white rounded-lg shadow p-4">
-            <div class="flex items-center mb-2">
-              <img src="https://via.placeholder.com/40" alt="프로필" class="rounded-full mr-2">
-              <div>
-                <p class="font-bold">박민수</p>
-                <p class="text-sm text-gray-500">1시간 전</p>
-              </div>
-            </div>
-            <p>주말에 등산 가실 분 계신가요? 함께 가요!</p>
-            <div class="mt-2 flex justify-between text-gray-500">
-              <button>좋아요</button>
-              <button>댓글</button>
-              <button>공유</button>
-            </div>
-          </div>
-
-          <div class="bg-white rounded-lg shadow p-4">
-            <div class="flex items-center mb-2">
-              <img src="https://via.placeholder.com/40" alt="프로필" class="rounded-full mr-2">
-              <div>
-                <p class="font-bold">정수연</p>
-                <p class="text-sm text-gray-500">2시간 전</p>
-              </div>
-            </div>
-            <p>새로 나온 영화 재미있대요. 같이 보러 갈 사람?</p>
-            <div class="mt-2 flex justify-between text-gray-500">
-              <button>좋아요</button>
-              <button>댓글</button>
-              <button>공유</button>
-            </div>
-          </div>
-        </div>
-      </main>
-
-      <footer class="bg-gray-200 p-4 text-center">
-        <p>&copy; 2024 항해플러스. All rights reserved.</p>
-      </footer>
-    </div>
-  </div>
-`;
+//renderMainPage
+const user = User();
+const renderMainPage = () => {
+  if (user.getUser()) {
+    document.querySelector('#root').innerHTML = LoginMain();
   } else {
-    document.querySelector('#root').innerHTML = `
-<div class="bg-gray-100 min-h-screen flex justify-center">
-    <div class="max-w-md w-full">
-      <header class="bg-blue-600 text-white p-4 sticky top-0">
-        <h1 class="text-2xl font-bold">항해플러스</h1>
-      </header>
-
-      <nav class="bg-white shadow-md p-2 sticky top-14">
-        <ul class="flex justify-around">
-          <li><a href="/" class="text-blue-600">홈</a></li>
-          <li><a href="/login" class="text-gray-600">로그인</a></li>
-        </ul>
-      </nav>
-
-      <main class="p-4">
-        <div class="mb-4 bg-white rounded-lg shadow p-4">
-          <textarea class="w-full p-2 border rounded" placeholder="무슨 생각을 하고 계신가요?"></textarea>
-          <button class="mt-2 bg-blue-600 text-white px-4 py-2 rounded">게시</button>
-        </div>
-
-        <div class="space-y-4">
-
-          <div class="bg-white rounded-lg shadow p-4">
-            <div class="flex items-center mb-2">
-              <img src="https://via.placeholder.com/40" alt="프로필" class="rounded-full mr-2">
-              <div>
-                <p class="font-bold">홍길동</p>
-                <p class="text-sm text-gray-500">5분 전</p>
-              </div>
-            </div>
-            <p>오늘 날씨가 정말 좋네요. 다들 좋은 하루 보내세요!</p>
-            <div class="mt-2 flex justify-between text-gray-500">
-              <button>좋아요</button>
-              <button>댓글</button>
-              <button>공유</button>
-            </div>
-          </div>
-
-          <div class="bg-white rounded-lg shadow p-4">
-            <div class="flex items-center mb-2">
-              <img src="https://via.placeholder.com/40" alt="프로필" class="rounded-full mr-2">
-              <div>
-                <p class="font-bold">김철수</p>
-                <p class="text-sm text-gray-500">15분 전</p>
-              </div>
-            </div>
-            <p>새로운 프로젝트를 시작했어요. 열심히 코딩 중입니다!</p>
-            <div class="mt-2 flex justify-between text-gray-500">
-              <button>좋아요</button>
-              <button>댓글</button>
-              <button>공유</button>
-            </div>
-          </div>
-
-          <div class="bg-white rounded-lg shadow p-4">
-            <div class="flex items-center mb-2">
-              <img src="https://via.placeholder.com/40" alt="프로필" class="rounded-full mr-2">
-              <div>
-                <p class="font-bold">이영희</p>
-                <p class="text-sm text-gray-500">30분 전</p>
-              </div>
-            </div>
-            <p>오늘 점심 메뉴 추천 받습니다. 뭐가 좋을까요?</p>
-            <div class="mt-2 flex justify-between text-gray-500">
-              <button>좋아요</button>
-              <button>댓글</button>
-              <button>공유</button>
-            </div>
-          </div>
-
-          <div class="bg-white rounded-lg shadow p-4">
-            <div class="flex items-center mb-2">
-              <img src="https://via.placeholder.com/40" alt="프로필" class="rounded-full mr-2">
-              <div>
-                <p class="font-bold">박민수</p>
-                <p class="text-sm text-gray-500">1시간 전</p>
-              </div>
-            </div>
-            <p>주말에 등산 가실 분 계신가요? 함께 가요!</p>
-            <div class="mt-2 flex justify-between text-gray-500">
-              <button>좋아요</button>
-              <button>댓글</button>
-              <button>공유</button>
-            </div>
-          </div>
-
-          <div class="bg-white rounded-lg shadow p-4">
-            <div class="flex items-center mb-2">
-              <img src="https://via.placeholder.com/40" alt="프로필" class="rounded-full mr-2">
-              <div>
-                <p class="font-bold">정수연</p>
-                <p class="text-sm text-gray-500">2시간 전</p>
-              </div>
-            </div>
-            <p>새로 나온 영화 재미있대요. 같이 보러 갈 사람?</p>
-            <div class="mt-2 flex justify-between text-gray-500">
-              <button>좋아요</button>
-              <button>댓글</button>
-              <button>공유</button>
-            </div>
-          </div>
-        </div>
-      </main>
-
-      <footer class="bg-gray-200 p-4 text-center">
-        <p>&copy; 2024 항해플러스. All rights reserved.</p>
-      </footer>
-    </div>
-  </div>
-`;
+    document.querySelector('#root').innerHTML = Main();
   }
 };
-
-function createRouter() {
-  const routes = {};
-
-  function addRoute(path, handler) {
-    routes[path] = handler;
-  }
-
-  function navigateTo(path) {
-    history.pushState(null, '', path);
-    handleRoute(path);
-  }
-
-  function handlePopState() {
-    handleRoute(window.location.pathname);
-  }
-
-  function handleRoute(path) {
-    const handler = routes[path];
-    if (handler) {
-      handler();
-    } else {
-      getErrorPage();
-    }
-  }
-
-  window.addEventListener('popstate', handlePopState);
-
-  return {
-    addRoute,
-    navigateTo,
-  };
-}
 
 const logout = () => {
-  localStorage.removeItem('user');
+  user.deleteUser();
   router.navigateTo('/login');
 };
-const getProfilePage = () => {
-  let user = JSON.parse(localStorage.getItem('user'));
-  if (!user) {
+const renderProfilePage = () => {
+  if (!user.getUser()) {
     history.replaceState(null, null, '/');
     return;
   }
 
-  document.querySelector('#root').innerHTML = `
-  <div class="bg-gray-100 min-h-screen flex justify-center">
-    <div class="max-w-md w-full">
-      <header class="bg-blue-600 text-white p-4 sticky top-0">
-        <h1 class="text-2xl font-bold">항해플러스</h1>
-      </header>
-      <nav class="bg-white shadow-md p-2 sticky top-14">
-        <ul class="flex justify-around">
-          <li><a href="/" class="text-gray-600">홈</a></li>
-          <li><a href="/profile" class="text-blue-600">프로필</a></li>
-          <li><a href="/logout" id="logout" class="text-gray-600">로그아웃</a></li>
-        </ul>
-      </nav>
-      <main class="p-4">
-        <div class="bg-white p-8 rounded-lg shadow-md">
-          <h2 class="text-2xl font-bold text-center text-blue-600 mb-8">내 프로필</h2>
-          <form id="profile-form">
-            <div class="mb-4">
-              <label for="username" class="block text-gray-700 text-sm font-bold mb-2">사용자 이름</label>
-              <input type="text" id="username" name="username" value="" class="w-full p-2 border rounded">
-            </div>
-            <div class="mb-4">
-              <label for="email" class="block text-gray-700 text-sm font-bold mb-2">이메일</label>
-              <input type="email" id="email" name="email" value="" class="w-full p-2 border rounded">
-            </div>
-            <div class="mb-6">
-              <label for="bio" class="block text-gray-700 text-sm font-bold mb-2">자기소개</label>
-              <textarea id="bio" name="bio" rows="4" class="w-full p-2 border rounded"></textarea>
-            </div>
-            <button type="submit" class="w-full bg-blue-600 text-white p-2 rounded font-bold">프로필 업데이트</button>
-          </form>
-        </div>
-      </main>
-      <footer class="bg-gray-200 p-4 text-center">
-        <p>&copy; 2024 항해플러스. All rights reserved.</p>
-      </footer>
-    </div>
-  </div>`;
-  user = JSON.parse(localStorage.getItem('user'));
-  if (user.username) {
-    document.querySelector('input[id="username"]').value = user.username;
+  document.querySelector('#root').innerHTML = Profile();
+  if (user.getUser().username) {
+    document.querySelector('input[id="username"]').value =
+      user.getUser().username;
   }
-  if (user.email) {
-    document.querySelector('input[id="email"]').value = user.email;
+  if (user.getUser().email) {
+    document.querySelector('input[id="email"]').value = user.getUser().email;
   }
-  if (user.bio) {
-    document.querySelector('textArea[id="bio"]').value = user.bio;
+  if (user.getUser().bio) {
+    document.querySelector('textArea[id="bio"]').value = user.getUser().bio;
   }
 
   const logoutButton = document.querySelector('a[id="logout"]');
@@ -325,7 +43,7 @@ const getProfilePage = () => {
     logout();
   });
 
-  const profileForm = document.getElementById('profile-form'); // form 요소를 가져옵니다.
+  const profileForm = document.getElementById('profile-form');
 
   const update = (username, email = '', bio = '') => {
     const updatedUser = {
@@ -333,7 +51,7 @@ const getProfilePage = () => {
       email: document.getElementById('email').value,
       bio: document.getElementById('bio').value,
     };
-    localStorage.setItem('user', JSON.stringify(updatedUser));
+    user.setUser(updatedUser);
     alert('프로필이 업데이트되었습니다.');
   };
 
@@ -351,37 +69,15 @@ const getProfilePage = () => {
       return;
     }
 
-    // 여기서 username과 함께 기본값을 사용하여 login 함수를 호출합니다.
     update(updatedUser);
   });
 };
 
-const getLoginPage = () => {
+const renderLoginPage = () => {
   console.log('Login Page');
-  document.querySelector('#root').innerHTML = `
-  <main class="bg-gray-100 flex items-center justify-center min-h-screen">
-      <div class="bg-white p-8 rounded-lg shadow-md w-full max-w-md">
-        <h1 class="text-2xl font-bold text-center text-blue-600 mb-8">항해플러스</h1>
-        <form id="login-form">
-          <div class="mb-4">
-            <input type="text" name="username" id="username" placeholder="이메일 또는 전화번호" class="w-full p-2 border rounded">
-          </div>
-          <div class="mb-6">
-            <input type="password" name="password" id="password" placeholder="비밀번호" class="w-full p-2 border rounded">
-          </div>
-          <button type="submit" class="w-full bg-blue-600 text-white p-2 rounded font-bold">로그인</button>
-        </form>
-        <div class="mt-4 text-center">
-          <a href="#" class="text-blue-600 text-sm">비밀번호를 잊으셨나요?</a>
-        </div>
-        <hr class="my-6">
-        <div class="text-center">
-          <button class="bg-green-500 text-white px-4 py-2 rounded font-bold">새 계정 만들기</button>
-        </div>
-      </div>
-    </main>`;
+  document.querySelector('#root').innerHTML = Login();
 
-  const loginForm = document.getElementById('login-form'); // form 요소를 가져옵니다.
+  const loginForm = document.getElementById('login-form');
 
   const login = (username, email = '', bio = '') => {
     let userInfo = {
@@ -390,12 +86,11 @@ const getLoginPage = () => {
       bio,
     };
 
-    localStorage.setItem('user', JSON.stringify(userInfo)); // user 정보를 localStorage에 저장합니다.
-    router.navigateTo('/profile'); // 로그인 후 프로필 페이지로 이동합니다.
+    user.setUser(userInfo);
+    router.navigateTo('/profile');
   };
 
   loginForm.addEventListener('submit', (e) => {
-    // form의 submit 이벤트를 사용합니다.
     e.preventDefault();
     console.log('login!');
     let username = document
@@ -406,35 +101,20 @@ const getLoginPage = () => {
       return;
     }
 
-    // 여기서 username과 함께 기본값을 사용하여 login 함수를 호출합니다.
     login(username);
   });
 };
 
-const getErrorPage = () => {
+const renderErrorPage = () => {
   console.log('Error Page');
-  document.querySelector('#root').innerHTML = `
-  <main class="bg-gray-100 flex items-center justify-center min-h-screen">
-      <div class="bg-white p-8 rounded-lg shadow-md w-full text-center" style="max-width: 480px">
-        <h1 class="text-2xl font-bold text-blue-600 mb-4">항해플러스</h1>
-        <p class="text-4xl font-bold text-gray-800 mb-4">404</p>
-        <p class="text-xl text-gray-600 mb-8">페이지를 찾을 수 없습니다</p>
-        <p class="text-gray-600 mb-8">
-          요청하신 페이지가 존재하지 않거나 이동되었을 수 있습니다.
-        </p>
-        <a href="/" class="bg-blue-600 text-white px-4 py-2 rounded font-bold">
-          홈으로 돌아가기
-        </a>
-      </div>
-    </main>`;
+  document.querySelector('#root').innerHTML = Error();
 };
-getMainPage();
-const router = createRouter();
-router.addRoute('/', getMainPage);
-router.addRoute('/login', getLoginPage);
-// router.addRoute('/logout', logout);
-router.addRoute('/profile', getProfilePage);
-router.addRoute('/404', getErrorPage);
+renderMainPage();
+const router = Router();
+router.addRoute('/', renderMainPage);
+router.addRoute('/login', renderLoginPage);
+router.addRoute('/profile', renderProfilePage);
+router.addRoute('/404', renderErrorPage);
 
 document.querySelector('nav').addEventListener('click', (e) => {
   if (e.target.tagName === 'A') {
